% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Weibull_v3.R
\name{SurvEffectWeibull}
\alias{SurvEffectWeibull}
\title{SurvEffectWeibull}
\usage{
SurvEffectWeibull(
  Data,
  t = 60,
  Treatment,
  SurvTime,
  Status,
  PS.formula,
  Censor.formula,
  Type = 1,
  Method = "IPTW",
  alpha = 0.1,
  q = 0.01
)
}
\arguments{
\item{q}{}
}
\description{
SurvEffectWeibull
}
