% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Cox_v2.R
\name{SurvEffectWithCox}
\alias{SurvEffectWithCox}
\title{SurvEffectWithCox}
\usage{
SurvEffectWithCox(
  Data,
  t = 60,
  Treatment,
  SurvTime,
  Status,
  PS.formula,
  Censor.formula,
  Type = 1,
  Method = "IPTW",
  alpha = 0.05,
  q = 0.01
)
}
\arguments{
\item{q}{}
}
\description{
SurvEffectWithCox
}
