% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalFun.R
\name{SurvFun}
\alias{SurvFun}
\title{Title}
\usage{
SurvFun(
  Data,
  t = 60,
  Treatment,
  SurvTime,
  Status,
  PS.formula,
  Censor.formula,
  Type = 1,
  Method = "IPTW"
)
}
\arguments{
\item{Method}{}
}
\description{
Title
}
